🐛[Fix] 상벌점 소수점이 0으로 절삭되던 문제 수정
PointLog.point 가 Int 인데 원본 ProfileChallengerPoint.point 는 Double(decodeDoubleFlexible)
이라, 매핑에서 Int(point.point) 로 절삭돼 -0.5 점이 조용히 0 이 됐다.

- PointLog.point 와 HomeGeneration 의 penaltyPoint/rewardPoint 를 Double 로 통일하고
  매핑의 Int 변환 제거. 합계도 Double 로 누적
- 절대 규칙 #2 는 "서버 응답 정수"의 String 통일 규칙이라 소수 값인 이 필드는 대상이 아니다.
  판단 근거를 PointLog 프로퍼티 주석에 남김
- 표시는 .number.precision(.fractionLength(0...1)) 로 2.0 → "2", -0.5 → "-0.5"
- 회귀 테스트 fetchMyProfileKeepsFractionalPoints 추가

+++ UMCApp/Features/Home/Domain/Sources/Models/PointLog.swift
@@ -15,13 +15,18 @@ public struct PointLog: Identifiable, Equatable, Sendable {
     public let reason: String
     /// 표시용 날짜 문자열 (예: "07.09")
     public let date: String
-    public let point: Int
+    /// 상벌점 값. 서버가 `-0.5` 같은 소수 배점을 내려주므로 `Double`로 유지한다.
+    ///
+    /// 절대 규칙 #2(서버 응답 "정수"는 전 레이어 `String`)의 대상이 아니다. 원본
+    /// ``ProfileChallengerPoint/point``가 `Double`이라 `String`으로 감싸면 합산·`abs`마다
+    /// 파싱이 필요하고, 실패 시 값이 사라진다. 손실 없이 그대로 옮기는 쪽이 단순하고 안전하다.
+    public let point: Double
     /// 상벌점 여부 (`true`: 상점, `false`: 벌점)
     public let isReward: Bool
 
     // MARK: - Init
 
-    public init(id: String, reason: String, date: String, point: Int, isReward: Bool) {
+    public init(id: String, reason: String, date: String, point: Double, isReward: Bool) {
         self.id = id
         self.reason = reason
         self.date = date
